+++ lib/repository/app_storage_repository/app_storage_repository.dart
@@ -1,10 +1,11 @@
+import "package:currency_converter/models/currency.dart";
 import "package:j1_theme/j1_theme.dart";
 
 abstract class AppStorageRepository extends J1ThemeRepository {
   // Favorites
-  Future<void> setFavorite(String code);
-  Future<void> removeFavorite(String code);
-  Stream<List<String>> getFavoritesStream();
+  Future<void> setFavorite(CurrencyCode code);
+  Future<void> removeFavorite(CurrencyCode code);
+  Stream<List<CurrencyCode>> getFavoritesStream();
 
   // Language
   Future<void> setLanguage(String languageCode);
