+++ lib/repository/app_storage_repository/realm/realm_text_theme.dart
@@ -1,7 +1,7 @@
 import "package:currency_converter/ui/themes/text_themes.dart";
 import "package:j1_theme/j1_theme.dart";
 import "package:realm/realm.dart";
-// ignore: unnecessary_import
+// ignore: unnecessary_import, depend_on_referenced_packages
 import "package:realm_common/realm_common.dart";
 
 part "realm_text_theme.realm.dart";
