+++ lib/repository/app_storage_repository/realm_catalog.dart
@@ -1,16 +1,8 @@
 import "package:realm/realm.dart";
-// ignore: unnecessary_import
+// ignore: unnecessary_import, depend_on_referenced_packages
 import "package:realm_common/realm_common.dart";
 
 part "realm_catalog.realm.dart";
-
-@RealmModel()
-class _RealmPageTransition {
-  @PrimaryKey()
-  late String key;
-
-  String pageTransition = "cupertino";
-}
 
 @RealmModel()
 class _RealmFavorites {
